+++ src/main/java/insane96mcp/enhancedai/modules/villager/VillagerAttacking.java
@@ -20,9 +20,9 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.Optional;
 
-@LoadFeature(module = Modules.Ids.VILLAGER, description = "Make villagers fight back. Use the enhancedai:villager_attacking/can_attack type tag to add more villagers, only works with entities that extend vanilla Villagers. Attack damage can't be changed via config due to limitation so use mods like Mobs Properties Randomness to change the damage. Base damage is 4.")
+@LoadFeature(module = Modules.Ids.VILLAGER, description = "Make villagers fight back. Use the enhancedai:villager/can_attack type tag to add more villagers, only works with entities that extend vanilla Villagers. Attack damage can't be changed via config due to limitation so use mods like Mobs Properties Randomness to change the damage. Base damage is 4.")
 public class VillagerAttacking extends Feature {
-    public static final TagKey<EntityType<?>> VILLAGERS_CAN_ATTACK = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("villager_attacking/can_attack"));
+    public static final TagKey<EntityType<?>> AFFECTED_ENTITY_TYPES = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("villager/can_attack"));
     @Config(description = "If true, villagers will attack back monsters")
     public static Boolean villagersFightBackEnemies = false;
     @Config(description = "Villagers will only attack players that have below this reputation (like Iron Golems by default). https://minecraft.wiki/w/Villager#Gossiping")
@@ -39,7 +39,7 @@ public class VillagerAttacking extends Feature {
 		FIGHTS_BACK_ENEMIES = EAIData.ofBool(this.createDataKey("fights_back_enemies"));
 		ATTACK_BELOW_REPUTATION = EAIData.ofInt(this.createDataKey("attack_below_reputation"));
 		SPEED_MULTIPLIER = EAIData.ofDouble(this.createDataKey("speed_multiplier"), (mob, speedMultiplier) -> {
-			if (!mob.getType().is(VILLAGERS_CAN_ATTACK))
+			if (!mob.getType().is(AFFECTED_ENTITY_TYPES))
 				return;
 			Optional<MeleeAttackGoal> meleeAttackGoal = GoalHelper.getGoal(mob.goalSelector, MeleeAttackGoal.class);
 			meleeAttackGoal.ifPresent(attackGoal -> ((MeleeAttackGoalAccessor) attackGoal).setSpeedModifier(speedMultiplier));
@@ -51,7 +51,7 @@ public class VillagerAttacking extends Feature {
     public void onSpawn(EntityJoinLevelEvent event) {
         if (!this.isEnabled()
                 || !(event.getEntity() instanceof Villager villager)
-                || !villager.getType().is(VILLAGERS_CAN_ATTACK))
+                || !villager.getType().is(AFFECTED_ENTITY_TYPES))
             return;
 
         villager.targetSelector.addGoal(1, (new EAIVillagerHurtByTargetGoal(villager)).setAlertOthers());